Adiciona rota para registrar nfce manualmente

+++ src/main/java/com/prati/projetomercado/dto/request/NfceDataRequest.java
@@ -1,5 +1,6 @@
 package com.prati.projetomercado.dto.request;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 
@@ -9,6 +10,6 @@ public record NfceDataRequest(
         AddressRequest address,
         String accessKey,
         LocalDate date,
-        double totalPrice,
+        BigDecimal totalPrice,
         List<ProductRequest> products
 ) {}
